requirements.txt
aiogram==2.25.1
mwsqlite==1.3.0
PyYAML==6.0.1